requirements.txt
streamlit==1.18.0
Pillow==9.0.1
pytesseract==0.3.8
pdfplumber==0.7.4
multiprocess==0.70.13
    